requirements.txt
PyQt5==5.10.1
sip==4.19.8
PyOpenGL
matplotlib
numpy
pyparsing
pyserial
scipy
